requirements.txt
# Core ML libraries
torch>=2.0.0
torchvision
torchaudio

# Hugging Face ecosystem
transformers>=4.40.0
diffusers>=0.28.0
accelerate>=0.25.0
datasets>=2.17.0
safetensors

# Video processing
opencv-python
imageio[ffmpeg]
av
decord

# Image processing
Pillow>=9.5.0
scikit-image

# Data handling
numpy
pandas
jsonlines

# GUI framework
gradio>=4.0.0

# Progress and logging
tqdm
wandb
tensorboard

# File handling
pathlib

# Configuration management
pyyaml
omegaconf

# Utility libraries
psutil
